// frontend/src/app/modules/candidate/interfaces/IGiftProgram.ts
export interface IGiftProgram {
  candidateName: string;
  currentDesignation: string;
  desiredDesignation: string;
  gapsIdentified: string;
  planOfAction: string;
  targetAchievementDate: string;
  isSupportRequired: boolean;
  endResult: string;
  majorAchievements: string;
  learningAndTransistionProcess: string;
  careerGrowthContribution: string;
  workRelatedTrainingAndCertifications: string;
  giftformAdminStatus: number;
  giftformManagerStatus: number;
  giftformRejectionReason: string;
}